pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "gauss-agent"
version = "0.2.2"
description = "Gauss is a focused Lean autoformalization workspace with a managed Claude Code launcher."
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Math Inc." }]
license = { text = "MIT" }
dependencies = [
  # Core
  "openai",
  "anthropic>=0.39.0",
  "python-dotenv",
  "fire",
  "httpx",
  "rich",
  "tenacity",
  "pyyaml",
  "requests",
  "jinja2",
  "pydantic>=2.0",
  # Interactive CLI (prompt_toolkit is used directly by cli.py)
  "prompt_toolkit",
]

[project.optional-dependencies]
modal = ["swe-rex[modal]>=1.4.0"]
daytona = ["daytona>=0.148.0"]
dev = ["pytest", "pytest-asyncio", "pytest-xdist", "mcp>=1.2.0"]
messaging = ["python-telegram-bot>=20.0", "discord.py[voice]>=2.0", "aiohttp>=3.9.0", "slack-bolt>=1.18.0", "slack-sdk>=3.27.0"]
cron = ["croniter"]
slack = ["slack-bolt>=1.18.0", "slack-sdk>=3.27.0"]
cli = ["simple-term-menu"]
lite = ["gauss-agent[cli]", "gauss-agent[pty]"]
gauss = ["gauss-agent[lite]", "gauss-agent[web]"]
web = ["firecrawl-py"]
image = ["fal-client"]
skills-hub = ["PyJWT[crypto]"]
speech = ["edge-tts", "faster-whisper>=1.0.0"]
full = [
  "gauss-agent[lite]",
  "gauss-agent[web]",
  "gauss-agent[image]",
  "gauss-agent[skills-hub]",
  "gauss-agent[speech]",
]
tts-premium = ["elevenlabs"]
voice = ["sounddevice>=0.4.6", "numpy>=1.24.0"]
pty = [
  "ptyprocess>=0.7.0; sys_platform != 'win32'",
  "pywinpty>=2.0.0; sys_platform == 'win32'",
]
mcp = ["mcp>=1.2.0"]
homeassistant = ["aiohttp>=3.9.0"]
acp = ["agent-client-protocol>=0.8.1,<1.0"]
rl = [
  "atroposlib @ git+https://github.com/NousResearch/atropos.git",
  "tinker @ git+https://github.com/thinking-machines-lab/tinker.git",
  "fastapi>=0.104.0",
  "uvicorn[standard]>=0.24.0",
  "wandb>=0.15.0",
]
yc-bench = ["yc-bench @ git+https://github.com/collinear-ai/yc-bench.git"]
all = [
  "gauss-agent[modal]",
  "gauss-agent[daytona]",
  "gauss-agent[messaging]",
  "gauss-agent[cron]",
  "gauss-agent[full]",
  "gauss-agent[dev]",
  "gauss-agent[tts-premium]",
  "gauss-agent[slack]",
  "gauss-agent[mcp]",
  "gauss-agent[homeassistant]",
  "gauss-agent[acp]",
  "gauss-agent[voice]",
]

[project.scripts]
gauss = "gauss_cli.main:main"
gauss-agent = "run_agent:main"
gauss-acp = "acp_adapter.entry:main"

[tool.setuptools]
py-modules = ["run_agent", "model_tools", "toolsets", "batch_runner", "trajectory_compressor", "toolset_distributions", "cli", "gauss_constants", "gauss_state", "gauss_time", "mini_swe_runner", "rl_cli", "utils"]

[tool.setuptools.packages.find]
include = ["agent", "tools", "tools.*", "gauss_cli", "gateway", "gateway.*", "cron", "acp_adapter"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: marks tests requiring external services (API keys, Modal, etc.)",
]
addopts = "-m 'not integration' -n auto"
